Allow the same project to live in separate containers
Having the container name exactly as the project name in the
`.nomad.yml` file is kind of limiting because you can't have the project
live in multiple containers (to try an experimental feature, for
example).

With this commit, we set a `user.nomad.homedir` config at container
creation to the absolute path of our project (config file). When we try
to run `nomad up` from a directory that isn't referenced by any
container, we create a new one and mangle its name if there's a naming
conflict.