Use a creation counter to ensure container name consistence
When creating containers without speciying their names, we should have a mechanism allowing us to identify each container before it is actually created using LXD. This
is why a creation counter has been added to the Container class. This attribute (that is in the namespace of the Container class) will be used to generate container
names when they are not provided by the .nomad.yml config file.