shell: add a way to set environment variables
Add a new config dictionary in the `shell` key: `environment`. All
values in this dictionary are passed to LXD's config under the
"environment.*" config type.

We set this config when `shell` is called, which is a bit at odds with
the rest of the container config, which is usually set at creation time.
We do this so that a user who fiddles with her `lxdock.yml` file doesn't
have to restart/recreate her container to benefit from the change she's
just made. They will be effective on the next `lxdock shell` call.