Add global provisioning capabilities
`provisioning` configuration items declared at the root are now applied
"globally" for provisioners that support it (only `ansible` for now).

Implementation of this feature required quite a refactoring to move the
provisioning logic around. In a nutshell: Less provisioning logic in
`Container` and more in `Provisioner` and `Project`.

One tricky matter was "when can we consider that a specific container
can have its `is_provisioned` flag set?". See code for answer.

I had to change quite many tests, so I used this opportunity to improve
tooling a bit to reduce code duplication in our tests, hence the
introduction of `FakeContainer`. I've only minimally applied its usage,
I'm sure this concept can be expanded and used elsewhere.

ref #88