Use Ubuntu 16.04 for Travis instead of 14.04
It seems that the Ubuntu 14.04 LXD repository has dissapeared.

Ubuntu 14.04 is getting quite old now, it's 4 years old and never shipped with LXD in the first place so I am unusure why we use it to run the tests.

So it seems time to set the OS for testing to 16.04 and phase out support for Trusty.