Go back to 14.04 but with a Snap
It turns out the Snap version of LXD is now available on 14.04

https://insights.ubuntu.com/2017/03/30/snaps-are-now-available-for-ubuntu-14-04-lts-desktop-and-server/

Based on work done in the conjure-up repository, thanks to battlemidget for the advice, as well as shuhaowu.

Moved to Makefile since we already have a Makefile and I wasn't sure how to do a while loop in the .travis.yml but do know how to do this in a Makefile.