Use older version of py.test
Travis is not liking uninstalling six in order to upgrade to a newew version.

py.test >= 3.3 requires a newer version of six, earlier versions do not seem to so pin py.test in a range

I don't know if this is because it isn't using a proper virtualenv when using sudo -E because this is what triggered this.