Split out flake8 and isort and move to test requirements
Because these are needed for running the tests but Sphinx is not, if we move these to test requirements Travis can avoid installing Sphinx.

Sphinx is not needed for the tests on Travis, but flake8 and isort are so should be moved to test requirements.

Update the Makefile so that make travis installs requirements-tests.txt instead of requirements-dev.txt

Version pin can be removed because docs can be built on any Xenial machine without problems.